backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
mangum
playwright
beautifulsoup4
